requirements.txt
python-telegram-bot==22.0
pytz==2024.1 